attempt to make deleteProduct work (still need to remove join table reference and delete associated purchase)

+++ instant_pot/src/main/java/com/web_services/instant_pot/domain/product/ProductLogic.java
@@ -31,7 +31,7 @@ public class ProductLogic{
 	    return pd.createProduct(partner, productName, productDescription, cost);
 	}
 	
-	public Product deleteProduct(int id){
+	public Product deleteProduct(Long id){
 		ProductDAL pd = new ProductDAL();
 		return pd.deleteProduct(id);
 	}
